requirements.txt
lantern-client
python-dotenv
parea-ai
asyncpg
seaborn